Created the MVVM skeleton.
FragmentHistory.kt is ready and connected to HistoryViewModel.kt and HistoryRepository.kt but need testing once we will have data.

fixed a bug in the Room DB and set an init method in MyApp.kt

Re made the FragmentLoginAndRegister.kt and set the google login from to work with the navigation graph.

added LiveData dependencies and implemented the LiveData in HistoryViewModel.kt, HistoryRepository.kt and ExerciseSummaryDao.kt